+++ tests/all.nim
@@ -13,6 +13,7 @@ import ./test_retry_allowance
 
 import ./test_ratelimit_token_bucket
 import ./test_ratelimit_fixed_window
+import ./test_ratelimit_gcra
 import ./test_ratelimit_keyed
 import ./test_ratelimit_results
 import ./test_ratelimit_sliding_window